backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.1